Battery Reports V0.1
This is a Batch script that Automatically Generates Battery Reports for any laptop that runs Windows 10, For a period of 3 days

Installation Steps: Download the Zip file and extract the Contents to C:\
Refer the Snapshot for Detailed Path to install

Execution : Double click on RUN this file Shortcut and Provide Admin Credentials If asked

Output: The Files are generated in .html format in output folder. Microsoft Edge Can be used to open the Reports for Better experience

Tell me how do You find my scripts, Provide your Valuable Suggestions and Comments on how the scripts can be improvised.